+++ src/main/java/com/jux/familyspace/component/MemberElementsSizeTracker.java
@@ -5,20 +5,21 @@ import com.jux.familyspace.model.FamilyMember;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 @Component
 public class MemberElementsSizeTracker implements ElementSizeTrackerInterface<FamilyMember> {
 
-    private final AtomicInteger totalElementsSize = new AtomicInteger();
+    private final AtomicLong totalElementsSize = new AtomicLong();
 
 
     @Override
-    public int getTotalSize() {
+    public long getTotalSize() {
         return totalElementsSize.get();
     }
 
     @Override
-    public void setTotalSize(int size) {
+    public void setTotalSize(long size) {
         totalElementsSize.set(size);
     }
